requirements.txt
uvicorn
fastapi
aiohttp
jsonpatch
pyyaml
jinja2